feat(schema) halt loading the plugin upon schema errors
This causes the loading of a plugin to early-exit in case of a schema error
instead of just providing a warning (and possibly failing in case the schema
table was corrupt beyond possible usage).

Also changes the behavior so that all plugins are attempted even after one
error, collecting all errors into a single error message.